constraints summary table
